requirements.txt
absl-py==2.4.0
adblockparser==0.7
aistudio-sdk==0.3.8
albucore==0.0.24
albumentations==2.0.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.9.0
astor==0.8.1
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
asyncpg==0.30.0
attrdict==2.0.1
autocommand==2.2.2
babel==2.18.0
backports.tarfile==1.2.0
bce-python-sdk==0.9.63
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==7.0.5
capybara-docsaid==0.12.0
certifi==2026.1.4
cffi==2.0.0
chardet==7.0.1
charset-normalizer==3.4.3
cheroot==10.0.1
CherryPy==18.10.0
cherrypy-cors==1.7.0
click==8.1.8
colorama==0.4.6
colored==2.3.1
coloredlogs==15.0.1
colorlog==6.10.1
comm==0.2.2
contourpy==1.3.2
cryptography==46.0.5
cssselect==1.4.0
cssutils==2.11.1
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
cycler==0.12.1
Cython==3.2.4
dacite==1.9.2
debugpy==1.8.14
decorator==5.2.1
dill==0.4.1
distlib==0.4.0
dnspython==2.7.0
easyocr==1.7.2
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
ExifRead==2.3.2
fastapi==0.115.12
filelock==3.19.1
fire==0.7.1
Flask==3.1.3
flask-babel==4.0.0
flatbuffers==25.9.23
fonttools==4.61.0
fsspec==2025.12.0
future==1.0.0
gast==0.7.0
gdown==5.2.0
google-pasta==0.2.0
greenlet==3.2.1
grpcio==1.78.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpagentparser==1.9.5
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
humanfriendly==10.0
idna==3.10
ImageIO==2.37.2
imagesize==2.0.0
imgaug==0.4.0
imgviz==1.7.6
ipaddr==2.2.0
ipwhois==1.1.0
ipykernel==6.29.5
ipython==8.37.0
itsdangerous==2.2.0
jaraco.collections==5.2.1
jaraco.context==6.0.1
jaraco.functools==4.2.1
jaraco.text==4.0.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==3.12.1
kiwisolver==1.4.9
labelme==5.10.1
lap==0.5.12
lazy_loader==0.4
libclang==18.1.1
lmdb==1.7.5
loguru==0.7.3
lxml==4.9.1
Mako==1.3.10
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.7
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.4
modelscope==1.34.0
more-itertools==10.8.0
mpmath==1.3.0
mrzscanner_docsaid==1.0.7
namex==0.1.0
natsort==8.4.0
nest-asyncio==1.6.0
netaddr==0.10.1
networkx==3.4.2
ninja==1.13.0
numpy==2.2.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.4.127
onnx==1.20.1
onnxruntime==1.23.2
onnxruntime-gpu==1.22.0
onnxslim==0.1.87
opencv-contrib-python==4.10.0.84
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
openpyxl==3.1.5
openvino==2025.4.1
openvino-telemetry==2025.2.0
opt-einsum==3.3.0
optree==0.18.0
osam==0.2.5
packaging==25.0
paddlepaddle-gpu==2.6.2
paddlex==3.4.2
pandas==2.3.3
parso==0.8.4
pdf2docx==0.5.8
pdf2image==1.17.0
phonenumbers==8.13.55
piexif==1.1.3
pillow==12.0.0
pillow_heif==1.3.0
pipenv==2025.0.4
platformdirs==4.3.8
polars==1.38.1
polars-runtime-32==1.38.1
portend==3.2.1
premailer==3.10.0
prettytable==3.17.0
prompt_toolkit==3.0.51
protobuf==6.33.5
psutil==7.0.0
psycopg2-binary==2.9.10
publicsuffixlist==0.9.4
pure_eval==0.2.3
py-cpuinfo==9.0.0
pybase64==1.4.3
pyclipper==1.4.0
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pygexf==0.2.2
Pygments==2.19.1
PyMuPDF==1.20.2
pyOpenSSL==21.0.0
pyparsing==3.2.5
PyPDF2==1.28.6
pypdfium2==5.6.0
PyQt5==5.15.9
PyQt5-Qt5==5.15.17
PyQt5-sip==12.11.0
PySocks==1.7.1
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==0.8.11
python-pptx==0.6.23
python-whois==0.7.3
PyTurboJPEG==1.8.3
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
RapidFuzz==3.14.3
rarfile==4.2
regex==2026.1.15
requests==2.32.4
rich==14.3.2
ruamel.yaml==0.19.1
ruff==0.15.5
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
secure==0.3.0
shapely==2.1.2
shellingham==1.5.4
simsimd==6.5.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
stack-data==0.6.3
starlette==0.46.2
stringzilla==4.2.3
sympy==1.13.1
tempora==5.8.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tokenizers==0.22.2
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchvision==0.21.0+cu124
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==5.1.0
triton==3.2.0
typer==0.23.0
typer-slim==0.23.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
ultralytics==8.4.14
ultralytics-thop==2.0.18
urllib3==1.26.20
uvicorn==0.34.2
virtualenv==20.34.0
visualdl==2.5.3
wcwidth==0.2.13
Werkzeug==3.1.5
wrapt==2.1.1
xlsxwriter==3.2.5
zc.lockfile==3.0.post1
